# specs/002-prd-json-cn/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Online Developer Tools Platform API
  description: API for the online developer tools platform providing JSON processing, code formatting, and execution capabilities
  version: 1.0.0
  contact:
    name: Platform Team
  license:
    name: MIT
servers:
  - url: https://api.{domain}/v1
    description: Production API
  - url: https://api-staging.{domain}/v1
    description: Staging API
paths:
  /tools:
    get:
      summary: Get available tools
      description: Retrieve list of available tools with categories and configurations
      tags:
        - Tools
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [json, formatting, execution, text, image, network, crypto]
          description: Filter tools by category
        - name: enabled_only
          in: query
          schema:
            type: boolean
            default: true
          description: Return only enabled tools
      responses:
        '200':
          description: List of available tools
          content:
            application/json:
              schema:
                type: object
                properties:
                  tools:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tool'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tools/{toolId}:
    get:
      summary: Get tool details
      description: Retrieve detailed information about a specific tool
      tags:
        - Tools
      parameters:
        - name: toolId
          in: path
          required: true
          schema:
            type: string
          description: Tool identifier
      responses:
        '200':
          description: Tool details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tool'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tools/json/format:
    post:
      summary: Format JSON
      description: Format and validate JSON data
      tags:
        - JSON Tools
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                json:
                  type: string
                  description: JSON string to format
                indent:
                  type: integer
                  default: 2
                  minimum: 0
                  maximum: 8
                  description: Number of spaces for indentation
                sort_keys:
                  type: boolean
                  default: false
                  description: Sort object keys alphabetically
              required:
                - json
      responses:
        '200':
          description: Formatted JSON
          content:
            application/json:
              schema:
                type: object
                properties:
                  formatted:
                    type: string
                    description: Formatted JSON string
                  valid:
                    type: boolean
                    description: Whether input was valid JSON
                  size:
                    type: integer
                    description: Input size in bytes
                  errors:
                    type: array
                    items:
                      type: string
                    description: Validation errors if any
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tools/json/validate:
    post:
      summary: Validate JSON
      description: Validate JSON data against schema
      tags:
        - JSON Tools
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                json:
                  type: string
                  description: JSON string to validate
                schema:
                  type: object
                  description: JSON Schema for validation
              required:
                - json
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    description: Validation result
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: string
                        message:
                          type: string
                    description: Validation errors
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tools/json/convert:
    post:
      summary: Convert JSON
      description: Convert JSON to other formats
      tags:
        - JSON Tools
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                json:
                  type: string
                  description: JSON string to convert
                target_format:
                  type: string
                  enum: [xml, yaml, csv, toml, properties]
                  description: Target format
                options:
                  type: object
                  description: Format-specific options
              required:
                - json
                - target_format
      responses:
        '200':
          description: Converted data
          content:
            application/json:
              schema:
                type: object
                properties:
                  converted:
                    type: string
                    description: Converted data
                  format:
                    type: string
                    description: Target format
                  size:
                    type: integer
                    description: Output size in bytes
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tools/code/format:
    post:
      summary: Format code
      description: Format source code in various languages
      tags:
        - Code Tools
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  description: Source code to format
                language:
                  type: string
                  enum: [javascript, typescript, css, html, sql, json, xml, yaml]
                  description: Programming language
                options:
                  type: object
                  description: Language-specific formatting options
              required:
                - code
                - language
      responses:
        '200':
          description: Formatted code
          content:
            application/json:
              schema:
                type: object
                properties:
                  formatted:
                    type: string
                    description: Formatted code
                  language:
                    type: string
                    description: Programming language
                  size:
                    type: integer
                    description: Output size in bytes
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /tools/code/execute:
    post:
      summary: Execute code
      description: Execute code in a secure WASM sandbox
      tags:
        - Code Execution
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  description: Source code to execute
                language:
                  type: string
                  enum: [javascript, typescript, python]
                  description: Programming language
                input:
                  type: string
                  description: Standard input for the program
                timeout:
                  type: integer
                  default: 5000
                  maximum: 30000
                  description: Execution timeout in milliseconds
              required:
                - code
                - language
      responses:
        '200':
          description: Execution result
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
                    description: Program output
                  error:
                    type: string
                    description: Error output if any
                  exit_code:
                    type: integer
                    description: Program exit code
                  execution_time:
                    type: integer
                    description: Execution time in milliseconds
                  memory_usage:
                    type: integer
                    description: Memory usage in bytes
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    Tool:
      type: object
      properties:
        id:
          type: string
          description: Tool identifier
        slug:
          type: string
          description: URL-friendly identifier
        name:
          type: string
          description: Display name
        category:
          type: string
          enum: [json, formatting, execution, text, image, network, crypto]
          description: Tool category
        description:
          type: string
          description: Tool description
        config:
          type: object
          description: Tool configuration
        enabled:
          type: boolean
          description: Whether tool is enabled
        beta:
          type: boolean
          description: Whether tool is in beta
        created_at:
          type: integer
          description: Creation timestamp
        updated_at:
          type: integer
          description: Last update timestamp
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error identifier
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
        timestamp:
          type: integer
          description: Error timestamp
        request_id:
          type: string
          description: Request identifier for debugging
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for tool access

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authenticated users
...